# Optimize maxStringLength