Because StringHolders store the strings as char array, using that directly instead of converting to a string saves a lot of time (and heap allocations).

Prior to this change Table.testReadInvalidCRA() ran in 117,394 ms with 72,509 ms being spent in StringArray.maxStringLength

With this change total time is 38,999 ms with StringArray.maxStringLength at 41.2 ms. There were also speed ups outside of maxStringLength which I believe is due to reducing GC.